recommendation_system_ml/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
scikit-surprise
textblob
nltk
joblib
networkx # if needed by Surprise
